requirements.txt
colorama
opencv-python
opencv-contrib-python
pillow_heif
Pillow